+++ src/com/vangav/backend/public_apis/car2go/json/operation_area/PlaceMark.java
@@ -50,7 +50,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class PlaceMark {
 
   @JsonProperty
-  public double[] coordinates;
+  public float[] coordinates;
   @JsonProperty
   public String name;
   @JsonProperty
