+++ src/com/vangav/backend/public_apis/car2go/json/vehicles/PlaceMark.java
@@ -52,7 +52,7 @@ public class PlaceMark {
   @JsonProperty
   public String address;
   @JsonProperty
-  public double[] coordinates;
+  public float[] coordinates;
   @JsonProperty
   public String exterior;
   @JsonProperty
@@ -74,7 +74,7 @@ public class PlaceMark {
    * @throws Exception
    */
   @JsonIgnore
-  public double getLongitude () throws Exception {
+  public float getLongitude () throws Exception {
     
     return this.coordinates[0];
   }
@@ -85,7 +85,7 @@ public class PlaceMark {
    * @throws Exception
    */
   @JsonIgnore
-  public double getLatitude () throws Exception {
+  public float getLatitude () throws Exception {
     
     return this.coordinates[1];
   }
@@ -96,7 +96,7 @@ public class PlaceMark {
    * @throws Exception
    */
   @JsonIgnore
-  public double getAltitude () throws Exception {
+  public float getAltitude () throws Exception {
     
     return this.coordinates[2];
   }
